+++ apps/web/src/app/admin-console/organizations/layouts/organization-layout.component.html
@@ -40,9 +40,9 @@
       ></bit-nav-item>
     </bit-nav-group>
     <bit-nav-item
-      *ngIf="isAccessIntelligenceFeatureEnabled"
-      [text]="'accessIntelligence' | i18n"
-      route="access-intelligence"
+      *ngIf="isRiskInsightsFeatureEnabled"
+      [text]="'RiskInsights' | i18n"
+      route="risk-insights"
     ></bit-nav-item>
     <bit-nav-group
       icon="bwi-billing"
