+++ apps/web/src/app/admin-console/organizations/layouts/organization-layout.component.ts
@@ -51,7 +51,7 @@ export class OrganizationLayoutComponent implements OnInit, OnDestroy {
   showPaymentAndHistory$: Observable<boolean>;
   hideNewOrgButton$: Observable<boolean>;
   organizationIsUnmanaged$: Observable<boolean>;
-  isAccessIntelligenceFeatureEnabled = false;
+  isRiskInsightsFeatureEnabled = false;
 
   private _destroy = new Subject<void>();
 
@@ -71,8 +71,8 @@ export class OrganizationLayoutComponent implements OnInit, OnDestroy {
   async ngOnInit() {
     document.body.classList.remove("layout_frontend");
 
-    this.isAccessIntelligenceFeatureEnabled = await this.configService.getFeatureFlag(
-      FeatureFlag.AccessIntelligence,
+    this.isRiskInsightsFeatureEnabled = await this.configService.getFeatureFlag(
+      FeatureFlag.RiskInsights,
     );
 
     this.organization$ = this.route.params
