--- apps/web/src/app/tools/access-intelligence/access-intelligence.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from "@angular/core";
-
-import { AccessIntelligenceRoutingModule } from "./access-intelligence-routing.module";
-import { AccessIntelligenceComponent } from "./access-intelligence.component";
-
-@NgModule({
-  imports: [AccessIntelligenceComponent, AccessIntelligenceRoutingModule],
-})
-export class AccessIntelligenceModule {}
