+++ apps/web/src/app/tools/risk-insights/all-applications.component.html
