+++ apps/web/src/app/tools/risk-insights/application-table.mock.ts
