+++ apps/web/src/app/tools/risk-insights/critical-applications.component.html
