+++ apps/web/src/app/tools/risk-insights/password-health-members-uri.component.spec.ts
@@ -1,43 +1,44 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ActivatedRoute, convertToParamMap } from "@angular/router";
-import { mock } from "jest-mock-extended";
+import { mock, MockProxy } from "jest-mock-extended";
 import { of } from "rxjs";
 
 // eslint-disable-next-line no-restricted-imports
-import { PasswordHealthService } from "@bitwarden/bit-common/tools/reports/access-intelligence";
+import { PasswordHealthService } from "@bitwarden/bit-common/tools/reports/risk-insights";
 import { AuditService } from "@bitwarden/common/abstractions/audit.service";
+import { OrganizationService } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { PasswordStrengthServiceAbstraction } from "@bitwarden/common/tools/password-strength";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { TableModule } from "@bitwarden/components";
-import { TableBodyDirective } from "@bitwarden/components/src/table/table.component";
 
 import { LooseComponentsModule } from "../../shared";
 import { PipesModule } from "../../vault/individual-vault/pipes/pipes.module";
 
-import { PasswordHealthComponent } from "./password-health.component";
+import { PasswordHealthMembersURIComponent } from "./password-health-members-uri.component";
 
-describe("PasswordHealthComponent", () => {
-  let component: PasswordHealthComponent;
-  let fixture: ComponentFixture<PasswordHealthComponent>;
+describe("PasswordHealthMembersUriComponent", () => {
+  let component: PasswordHealthMembersURIComponent;
+  let fixture: ComponentFixture<PasswordHealthMembersURIComponent>;
+  let cipherServiceMock: MockProxy<CipherService>;
+  const passwordHealthServiceMock = mock<PasswordHealthService>();
+
   const activeRouteParams = convertToParamMap({ organizationId: "orgId" });
 
   beforeEach(async () => {
+    cipherServiceMock = mock<CipherService>();
     await TestBed.configureTestingModule({
-      imports: [PasswordHealthComponent, PipesModule, TableModule, LooseComponentsModule],
-      declarations: [TableBodyDirective],
+      imports: [PasswordHealthMembersURIComponent, PipesModule, TableModule, LooseComponentsModule],
       providers: [
-        { provide: CipherService, useValue: mock<CipherService>() },
+        { provide: CipherService, useValue: cipherServiceMock },
         { provide: I18nService, useValue: mock<I18nService>() },
         { provide: AuditService, useValue: mock<AuditService>() },
+        { provide: OrganizationService, useValue: mock<OrganizationService>() },
         {
           provide: PasswordStrengthServiceAbstraction,
           useValue: mock<PasswordStrengthServiceAbstraction>(),
         },
-        {
-          provide: PasswordHealthService,
-          useValue: mock<PasswordHealthService>(),
-        },
+        { provide: PasswordHealthService, useValue: passwordHealthServiceMock },
         {
           provide: ActivatedRoute,
           useValue: {
@@ -50,15 +51,11 @@ describe("PasswordHealthComponent", () => {
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(PasswordHealthComponent);
+    fixture = TestBed.createComponent(PasswordHealthMembersURIComponent);
     component = fixture.componentInstance;
-
-    fixture.detectChanges();
   });
 
   it("should initialize component", () => {
     expect(component).toBeTruthy();
   });
-
-  it("should call generateReport on init", () => {});
 });
