+++ apps/web/src/app/tools/risk-insights/password-health-members-uri.component.ts
@@ -1,50 +1,54 @@
+import { CommonModule } from "@angular/common";
 import { Component, DestroyRef, inject, OnInit } from "@angular/core";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
-import { FormControl, FormsModule } from "@angular/forms";
 import { ActivatedRoute } from "@angular/router";
-import { debounceTime, map } from "rxjs";
+import { map } from "rxjs";
 
+import { JslibModule } from "@bitwarden/angular/jslib.module";
 // eslint-disable-next-line no-restricted-imports
-import { PasswordHealthService } from "@bitwarden/bit-common/tools/reports/access-intelligence";
+import { PasswordHealthService } from "@bitwarden/bit-common/tools/reports/risk-insights";
 import { AuditService } from "@bitwarden/common/abstractions/audit.service";
+import { OrganizationService } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
+import { Organization } from "@bitwarden/common/admin-console/models/domain/organization";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { PasswordStrengthServiceAbstraction } from "@bitwarden/common/tools/password-strength";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
 import {
+  BadgeModule,
   BadgeVariant,
-  SearchModule,
+  ContainerComponent,
   TableDataSource,
   TableModule,
-  ToastService,
 } from "@bitwarden/components";
-import { CardComponent } from "@bitwarden/tools-card";
 
+// eslint-disable-next-line no-restricted-imports
 import { HeaderModule } from "../../layouts/header/header.module";
 // eslint-disable-next-line no-restricted-imports
-import { SharedModule } from "../../shared";
 import { OrganizationBadgeModule } from "../../vault/individual-vault/organization-badge/organization-badge.module";
 // eslint-disable-next-line no-restricted-imports
 import { PipesModule } from "../../vault/individual-vault/pipes/pipes.module";
 
 @Component({
   standalone: true,
-  selector: "tools-password-health-members",
-  templateUrl: "password-health-members.component.html",
+  selector: "tools-password-health-members-uri",
+  templateUrl: "password-health-members-uri.component.html",
   imports: [
-    CardComponent,
+    BadgeModule,
     OrganizationBadgeModule,
+    CommonModule,
+    ContainerComponent,
     PipesModule,
+    JslibModule,
     HeaderModule,
-    SearchModule,
-    FormsModule,
-    SharedModule,
     TableModule,
   ],
   providers: [PasswordHealthService],
 })
-export class PasswordHealthMembersComponent implements OnInit {
+export class PasswordHealthMembersURIComponent implements OnInit {
   passwordStrengthMap = new Map<string, [string, BadgeVariant]>();
+
+  weakPasswordCiphers: CipherView[] = [];
 
   passwordUseMap = new Map<string, number>();
 
@@ -54,26 +58,23 @@ export class PasswordHealthMembersComponent implements OnInit {
 
   dataSource = new TableDataSource<CipherView>();
 
+  reportCiphers: (CipherView & { hostURI: string })[] = [];
+  reportCipherURIs: string[] = [];
+
+  organization: Organization;
+
   loading = true;
-
-  selectedIds: Set<number> = new Set<number>();
-
-  protected searchControl = new FormControl("", { nonNullable: true });
 
   private destroyRef = inject(DestroyRef);
 
   constructor(
     protected cipherService: CipherService,
     protected passwordStrengthService: PasswordStrengthServiceAbstraction,
+    protected organizationService: OrganizationService,
     protected auditService: AuditService,
     protected i18nService: I18nService,
     protected activatedRoute: ActivatedRoute,
-    protected toastService: ToastService,
-  ) {
-    this.searchControl.valueChanges
-      .pipe(debounceTime(200), takeUntilDestroyed())
-      .subscribe((v) => (this.dataSource.filter = v));
-  }
+  ) {}
 
   ngOnInit() {
     this.activatedRoute.paramMap
@@ -97,40 +98,11 @@ export class PasswordHealthMembersComponent implements OnInit {
 
     await passwordHealthService.generateReport();
 
-    this.dataSource.data = passwordHealthService.reportCiphers;
-
+    this.dataSource.data = passwordHealthService.groupCiphersByLoginUri();
     this.exposedPasswordMap = passwordHealthService.exposedPasswordMap;
     this.passwordStrengthMap = passwordHealthService.passwordStrengthMap;
     this.passwordUseMap = passwordHealthService.passwordUseMap;
     this.totalMembersMap = passwordHealthService.totalMembersMap;
     this.loading = false;
-  }
-
-  markAppsAsCritical = async () => {
-    // TODO: Send to API once implemented
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        this.selectedIds.clear();
-        this.toastService.showToast({
-          variant: "success",
-          title: null,
-          message: this.i18nService.t("appsMarkedAsCritical"),
-        });
-        resolve(true);
-      }, 1000);
-    });
-  };
-
-  trackByFunction(_: number, item: CipherView) {
-    return item.id;
-  }
-
-  onCheckboxChange(id: number, event: Event) {
-    const isChecked = (event.target as HTMLInputElement).checked;
-    if (isChecked) {
-      this.selectedIds.add(id);
-    } else {
-      this.selectedIds.delete(id);
-    }
   }
 }
