+++ apps/web/src/app/tools/risk-insights/password-health.component.spec.ts
@@ -1,44 +1,43 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ActivatedRoute, convertToParamMap } from "@angular/router";
-import { mock, MockProxy } from "jest-mock-extended";
+import { mock } from "jest-mock-extended";
 import { of } from "rxjs";
 
 // eslint-disable-next-line no-restricted-imports
-import { PasswordHealthService } from "@bitwarden/bit-common/tools/reports/access-intelligence";
+import { PasswordHealthService } from "@bitwarden/bit-common/tools/reports/risk-insights";
 import { AuditService } from "@bitwarden/common/abstractions/audit.service";
-import { OrganizationService } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { PasswordStrengthServiceAbstraction } from "@bitwarden/common/tools/password-strength";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { TableModule } from "@bitwarden/components";
+import { TableBodyDirective } from "@bitwarden/components/src/table/table.component";
 
 import { LooseComponentsModule } from "../../shared";
 import { PipesModule } from "../../vault/individual-vault/pipes/pipes.module";
 
-import { PasswordHealthMembersURIComponent } from "./password-health-members-uri.component";
+import { PasswordHealthComponent } from "./password-health.component";
 
-describe("PasswordHealthMembersUriComponent", () => {
-  let component: PasswordHealthMembersURIComponent;
-  let fixture: ComponentFixture<PasswordHealthMembersURIComponent>;
-  let cipherServiceMock: MockProxy<CipherService>;
-  const passwordHealthServiceMock = mock<PasswordHealthService>();
-
+describe("PasswordHealthComponent", () => {
+  let component: PasswordHealthComponent;
+  let fixture: ComponentFixture<PasswordHealthComponent>;
   const activeRouteParams = convertToParamMap({ organizationId: "orgId" });
 
   beforeEach(async () => {
-    cipherServiceMock = mock<CipherService>();
     await TestBed.configureTestingModule({
-      imports: [PasswordHealthMembersURIComponent, PipesModule, TableModule, LooseComponentsModule],
+      imports: [PasswordHealthComponent, PipesModule, TableModule, LooseComponentsModule],
+      declarations: [TableBodyDirective],
       providers: [
-        { provide: CipherService, useValue: cipherServiceMock },
+        { provide: CipherService, useValue: mock<CipherService>() },
         { provide: I18nService, useValue: mock<I18nService>() },
         { provide: AuditService, useValue: mock<AuditService>() },
-        { provide: OrganizationService, useValue: mock<OrganizationService>() },
         {
           provide: PasswordStrengthServiceAbstraction,
           useValue: mock<PasswordStrengthServiceAbstraction>(),
         },
-        { provide: PasswordHealthService, useValue: passwordHealthServiceMock },
+        {
+          provide: PasswordHealthService,
+          useValue: mock<PasswordHealthService>(),
+        },
         {
           provide: ActivatedRoute,
           useValue: {
@@ -51,11 +50,15 @@ describe("PasswordHealthMembersUriComponent", () => {
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(PasswordHealthMembersURIComponent);
+    fixture = TestBed.createComponent(PasswordHealthComponent);
     component = fixture.componentInstance;
+
+    fixture.detectChanges();
   });
 
   it("should initialize component", () => {
     expect(component).toBeTruthy();
   });
+
+  it("should call generateReport on init", () => {});
 });
