+++ apps/web/src/app/tools/risk-insights/risk-insights-routing.module.ts
@@ -4,15 +4,15 @@ import { RouterModule, Routes } from "@angular/router";
 import { canAccessFeature } from "@bitwarden/angular/platform/guard/feature-flag.guard";
 import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
 
-import { AccessIntelligenceComponent } from "./access-intelligence.component";
+import { RiskInsightsComponent } from "./risk-insights.component";
 
 const routes: Routes = [
   {
     path: "",
-    component: AccessIntelligenceComponent,
-    canActivate: [canAccessFeature(FeatureFlag.AccessIntelligence)],
+    component: RiskInsightsComponent,
+    canActivate: [canAccessFeature(FeatureFlag.RiskInsights)],
     data: {
-      titleId: "accessIntelligence",
+      titleId: "RiskInsights",
     },
   },
 ];
@@ -21,4 +21,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
-export class AccessIntelligenceRoutingModule {}
+export class RiskInsightsRoutingModule {}
