+++ apps/web/src/app/tools/risk-insights/risk-insights.component.html
@@ -1,4 +1,4 @@
-<div class="tw-mb-1 text-primary" bitTypography="body1">{{ "accessIntelligence" | i18n }}</div>
+<div class="tw-mb-1 text-primary" bitTypography="body1">{{ "riskInsights" | i18n }}</div>
 <h1 bitTypography="h1">{{ "passwordRisk" | i18n }}</h1>
 <div class="tw-text-muted">{{ "discoverAtRiskPasswords" | i18n }}</div>
 <div class="tw-bg-primary-100 tw-rounded-lg tw-w-full tw-px-8 tw-py-2 tw-my-4">
