+++ apps/web/src/app/tools/risk-insights/risk-insights.component.ts
@@ -15,7 +15,7 @@ import { PasswordHealthMembersURIComponent } from "./password-health-members-uri
 import { PasswordHealthMembersComponent } from "./password-health-members.component";
 import { PasswordHealthComponent } from "./password-health.component";
 
-export enum AccessIntelligenceTabType {
+export enum RiskInsightsTabType {
   AllApps = 0,
   CriticalApps = 1,
   NotifiedMembers = 2,
@@ -23,7 +23,7 @@ export enum AccessIntelligenceTabType {
 
 @Component({
   standalone: true,
-  templateUrl: "./access-intelligence.component.html",
+  templateUrl: "./risk-insights.component.html",
   imports: [
     AllApplicationsComponent,
     AsyncActionsModule,
@@ -39,8 +39,8 @@ export enum AccessIntelligenceTabType {
     TabsModule,
   ],
 })
-export class AccessIntelligenceComponent {
-  tabIndex: AccessIntelligenceTabType;
+export class RiskInsightsComponent {
+  tabIndex: RiskInsightsTabType;
   dataLastUpdated = new Date();
 
   apps: any[] = [];
@@ -70,7 +70,7 @@ export class AccessIntelligenceComponent {
     private router: Router,
   ) {
     route.queryParams.pipe(takeUntilDestroyed()).subscribe(({ tabIndex }) => {
-      this.tabIndex = !isNaN(tabIndex) ? tabIndex : AccessIntelligenceTabType.AllApps;
+      this.tabIndex = !isNaN(tabIndex) ? tabIndex : RiskInsightsTabType.AllApps;
     });
   }
 }
