+++ apps/web/src/app/tools/risk-insights/risk-insights.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from "@angular/core";
+
+import { RiskInsightsRoutingModule } from "./risk-insights-routing.module";
+import { RiskInsightsComponent } from "./risk-insights.component";
+
+@NgModule({
+  imports: [RiskInsightsComponent, RiskInsightsRoutingModule],
+})
+export class RiskInsightsModule {}
