+++ apps/web/src/locales/en/messages.json
@@ -5,8 +5,8 @@
   "criticalApplications": {
     "message": "Critical applications"
   },
-  "accessIntelligence": {
-    "message": "Access Intelligence"
+  "riskInsights": {
+    "message": "Risk Insights"
   },
   "passwordRisk": {
     "message": "Password Risk"
